apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pillow>=11.0.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
openai>=1.54.0
pytest>=8.3.0
